Basic Kanban features using localStorage Changes New /components directory with footer containing theme toggle (may add accent color selection), footer, and Kanban management component. Modified globals.css to remove extraneous styling that interferes with next-themes. Modified main layout.tsx to add theme provider with class attribute which gets toggled with the component and default system theme preference. Current Kanban features are basically just adding lists and then adding tasks to a list and everything persists between reloads so localStorage is set up properly. I have functions set up for removing tasks and also marking them as completed but need to set up ui components to do that. Footer just contains the theme toggle and lays it out where I want.
Future Directions
Going to make some small forms to render when adding a list or adding a task to a list and then some sort of dropdown to render delete option on tasks and lists (need function to delete an entire list). I want to make a collapsible sidebar to show boards and add/remove boards. I also want to switch from simply mapping the existing items to mapping the items into list items and using an ordered list so that I can change the order by dragging lists and tasks around (need functions to move tasks from one list to another and change the order of the lists) and figure out what to use to actually handle moving the items around. I then want to make a modal for editing tasks and possibly setup steps as a checklist like trello has. But what I really would like to do is get some CI/CD setup, some unit tests with jest and then end-to-end tests with cypress, set up some sort of deployment on Netlify just to see what that's like and then I also want to containerize the app with Docker to see how that is as well.

Note
N/A